pyproject.toml
[tool.poetry]
name = "api-dados-rio"
version = "0.1.0"
description = "API de dados abertos do Escritório de Dados."
authors = ["Escritório de Dados Rio <escritoriodedados@gmail.com>"]
license = "GPL-3.0-only"
readme = "README.md"
homepage = "https://github.com/prefeitura-rio/api-dados-rio"
repository = "https://github.com/prefeitura-rio/api-dados-rio"
keywords = ["python", "template"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
Django = "^4.1"
django-cors-headers = "^3.13.0"
djangorestframework = "^3.13.1"
django-filter = "^22.1"
gunicorn = "^20.1.0"
django-redis = "^5.2.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
black = "20.8b1"
pre-commit = "^2.18.1"
pytest-cov = "^3.0.0"
flake8 = "^4.0.1"
pdoc3 = "^0.10.0"

[tool.poetry.scripts]
docs = "scripts.docs:main"
lint = "scripts.lint:main"
test = "scripts.test:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
